Fix effects clear automation on start/end combat
